Fix streaming size issues

+++ inc/streaming.h
@@ -37,11 +37,11 @@ typedef struct
  ******************************************************************************/
 streaming_channel_t Stream_CreateStreamingChannel(const void *ring_buffer, uint16_t ring_buffer_size, uint8_t data_size);
 void Stream_ResetStreamingChannel(streaming_channel_t *stream);
-uint8_t Stream_PutSample(streaming_channel_t *stream, const void *data, uint16_t size);
-uint8_t Stream_GetSample(streaming_channel_t *stream, void *data, uint16_t size);
-uint8_t Stream_GetAvailableSampleNB(streaming_channel_t *stream);
-uint8_t Stream_GetAvailableSampleNBUntilEndBuffer(streaming_channel_t *stream);
-uint8_t Stream_AddAvailableSampleNB(streaming_channel_t *stream, uint16_t size);
-uint8_t Stream_RmvAvailableSampleNB(streaming_channel_t *stream, uint16_t size);
+uint16_t Stream_PutSample(streaming_channel_t *stream, const void *data, uint16_t size);
+uint16_t Stream_GetSample(streaming_channel_t *stream, void *data, uint16_t size);
+uint16_t Stream_GetAvailableSampleNB(streaming_channel_t *stream);
+uint16_t Stream_GetAvailableSampleNBUntilEndBuffer(streaming_channel_t *stream);
+uint16_t Stream_AddAvailableSampleNB(streaming_channel_t *stream, uint16_t size);
+uint16_t Stream_RmvAvailableSampleNB(streaming_channel_t *stream, uint16_t size);
 
 #endif /* LUOS_H */
